Rename genre_helper.rb to mock_helper.rb and modify label_spec.rb
