--- spec/label_mock.rb
@@ -1,7 +1,0 @@
-class Item
-  attr_accessor :label
-
-  def initialize(published_date)
-    @published_date = published_date
-  end
-end
